libvpx Chrome build on Windows
This change allows libvpx to build on Windows with custom build step to
generate assembly offsets.

Added a tool with this change:

- copy_obj.sh

This tool copies a file a list of candidate locations. Such a tool is needed
because there are multiple possible locations where a .obj file is located,
depending on MSVS or Ninja is used.

Review URL: https://chromiumcodereview.appspot.com/10459033

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/libvpx@139713 4ff67af0-8c30-449e-8e8b-ad334ec8d88c