Attempt to let libvpx's SSE files build with clang-cl.
cl.exe can use intrinsics even if the arch targeted by the compiler doesn't
support the SSE level needed by the intrinsic.

clang (and gcc) can't do this, because it converts intrinsics into general
LLVM operations, and the LLVM bitcode is then translated back into assembly
later on, based on the target arch.

So explicitly enable SSSE3 and SSE4 when using clang-cl, so that the intrinsics
compile.

(It's unfortunate that clang-cl deviates fairly heavily from cl in this regard.
But this is very difficult to change, and it's the only larger deviation so
far.)

Depends on https://codereview.chromium.org/508433003/

Similar to https://codereview.chromium.org/475273002/ and
https://codereview.chromium.org/493883002/

BUG=82385
Review URL: https://codereview.chromium.org/504893003/


git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/libvpx@291625 4ff67af0-8c30-449e-8e8b-ad334ec8d88c