Cherry pick final avx[2] detection code
https://gerrit.chromium.org/gerrit/73502
https://gerrit.chromium.org/gerrit/73581

The original implementation of the avx2 detection appeared to lock up
WinXP/Vista bots. Pull the final version which guards the xgetbv call with the
cpu flags for both 256-bit AVX *and* OSXSAVE.

Theory is that AVX was declared available by the cpuid call (bit 28) but the OS
did not support it (as would be indicated by OSXSAVE being unset). The check
was incomplete and allowed it to step forward to the next checks: xgetbv and
cpuid for leaf 7 feature flags. One of those (most likely xgetbv since the leaf
7 check has been there for a while) causes the process to hang and the tests to
time out. That's the theory anyway.

R=tomfinegan@chromium.org, jzern@chromium.org
CC=bmeurer@chromium.org

Review URL: https://codereview.chromium.org//869853003